Create static BitmapDrawer.
Add method to draw cross arount chosen point on the bitmap.
Integrate it into a SpecificPointsReductionForm.
Has some problems:
Unhandled exception if image is smaller than picture box and you try to
click on the part where there is no image.

If image is bigger, stretch does not help to get coordinates of the part
of the image that would normally be hidden. Maybe somehow to change
size, resolution of the image, but that would result in quality change.